% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_func.R
\name{cdf_dl}
\alias{cdf_dl}
\title{Calculate conditional CDFs}
\usage{
cdf_dl(mod, new.data, at.y = 0, se = TRUE)
}
\arguments{
\item{mod}{the multiple DLs model}

\item{new.data}{the new data}

\item{at.y}{P(y <= at.y | new.data)}

\item{se}{if confidence intervals needed (default = TRUE)}
}
\value{
A list of estimation, standard error, and confidence intervals
}
\description{
This functions calculates the conditional CDFs based on the model
and new data.
}
