% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_function.R
\name{func_link}
\alias{func_link}
\title{Link functions}
\usage{
func_link(link)
}
\arguments{
\item{link}{the link function}
}
\value{
A list of functions subject to a link function
}
\description{
This function includes necessary functions related to each link function
}
