% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_function.R
\name{func_link_num}
\alias{func_link_num}
\title{Link functions (number)}
\usage{
func_link_num(link)
}
\arguments{
\item{link}{the link function}
}
\value{
An integer representing corresponding link function
}
\description{
This function faciliates the stan code (used as an internal function)
}
