% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_V.R
\name{func_V}
\alias{func_V}
\title{Calculate the covariance matrix}
\usage{
func_V(coef, n, x, y, delta, k, p, fam)
}
\arguments{
\item{coef}{coefficients (alpha, beta)}

\item{n}{number of subjects}

\item{x}{original covariate matrix}

\item{y}{ranks of code values}

\item{delta}{censoring indicators}

\item{k}{the number of unique code values}

\item{p}{the number of covariates}

\item{fam}{a list of functions subject to the link function}
}
\value{
A covariance matrix of coefficients
}
\description{
This functions calculates the covariance matrix based on the point estimates
}
