% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleDL-package.R
\docType{package}
\name{multipleDL-package}
\alias{multipleDL-package}
\title{Address Detection Limits by Cumulative Probability Models}
\description{
The package allows fitting regression models on continuous/ordinal response data subject to detection limits (DLs)
based on cumulative probability models (CPMs). Both single and multiple DLs can be handled. Conditional quantiles and CDFs
(cumulative distribution functions) can obtained from fitted models.
}
\details{
The 'multipleDL' package.
}
\references{
Stan Development Team (2020). RSroxygen2::roxygenize()tan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
Harrell, F. (2020). rms: Regression modeling strategies. R package version 6.1.0. https://CRAN.R-project.org/package=rms
Tian et al. "Addressing detection limits by semiparametric cumulative probability models." (2022) (to be submitted)
}
