% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multpledls.R
\name{multipledls}
\alias{multipledls}
\title{CPMs for multiple DLs}
\usage{
multipledls(formula, data, delta_lower = NULL, delta_upper = NULL, link)
}
\arguments{
\item{formula}{a R formula object}

\item{data}{a data frame}

\item{delta_lower}{(optional) indicators of lower DLs censoring (1: observed; 0:censored). If not specified, treat as observed.}

\item{delta_upper}{(optional) indicators of upper DLs censoring(1: observed; 0:censored). If not specified, treat as observed.}

\item{link}{the link function}
}
\value{
A list of results: coefficients, covariance matrix, unique response values, number of alphas, number of betas, link functions
}
\description{
This function build the CPM for multiple DLs
}
