% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_func.R
\name{quantile_dl}
\alias{quantile_dl}
\title{Calculate conditional quantiles}
\usage{
quantile_dl(mod, new.data, probs = 0.5, se = TRUE)
}
\arguments{
\item{mod}{the multiple DLs model}

\item{new.data}{the new data}

\item{probs}{pth quantiles}

\item{se}{if confidence intervals needed (default = TRUE)}
}
\value{
A list of estimation and confidence intervals
}
\description{
This functions calculates the conditional weighted quantiles based on the model
and new data.
}
